spellcheck
this is a personal indictment against you
