+++ src/main/java/latibot/command/commands/audio/SpeakCmd.java
@@ -1,4 +1,4 @@
-package latibot.command.comands.audio;
+package latibot.command.commands.audio;
 
 import java.io.File;
 import java.io.FileInputStream;
